rename: profile-tabs fix
